Added a balanced binary search tree to find the genes from the peaks in garnet
The Binary search tree reduced the runtime for the search for the
mapping of each peak region to a gene. Initially we were gearching
through the gene file for each peak but now we only need to go through
the entire gene file once to construct the Balanced Binay Search tree,
after that it should only be a mater of searching this tree to find the
ones within peak regions.