added contributor to AVL tree implementation - garnet
This document detains the Balanced Binary Search Tree datastructure
which I later use in the map_peaks_to_genes.py script. It is what I use
to sort the genes so that we can effieciently search through it in O(log
n) time. I mave made some changes to it but most of the doument was
adapted from Pavel Grafov's github profile